+++ src/bin/lisp/with-prof.lisp
@@ -1,0 +1,16 @@
+(and (builtin? 'time)
+     (number? +bps+)
+  (macro with-prof body
+    (with ((stime    (symbol))
+           (duration (symbol)))
+      $(let ,stime (time)
+         (block nil
+           ,@body)
+         (let ,duration (- (time) ,stime)
+           (out "Time spent: ")
+           (print (/ ,duration +bps+))
+           (out 's)
+           (out " (")
+           (print ,duration)
+           (out " bekloppies)")
+           (terpri))))))